updated migration, models, and changed date input to date picker

+++ www/database/migrations/2019_10_08_082605_create_news_table.php
@@ -20,7 +20,8 @@ class CreateNewsTable extends Migration
             $table->string('date');
             $table->text('content');
             $table->string('image');
-            $table->unsignedInteger('user_id');
+            // $table->unsignedInteger('user_id');
+            $table->unsignedBigInteger('user_id');
             //$table->timestamp('updated_at')->nullable();
             //$table->timestamp('created_at')->nullable();
             $table->timestamps();
